# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Laravel API",
        "description": "這是我們的 API 文件",
        "contact": {
            "email": "xxx"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/createinventory": {
            "post": {
                "tags": [
                    "base_inventory"
                ],
                "summary": "新增庫別資訊",
                "description": "新增庫別資訊",
                "operationId": "createinventory",
                "parameters": [
                    {
                        "name": "inventory_no",
                        "in": "query",
                        "description": "庫別代號",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "inventory_nm",
                        "in": "query",
                        "description": "庫別名稱",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "inventory_qty",
                        "in": "query",
                        "description": "庫存數量",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "lot_num",
                        "in": "query",
                        "description": "批號",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "safety_stock",
                        "in": "query",
                        "description": "安全庫存",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "lastStock_receiptdate",
                        "in": "query",
                        "description": "最近一次進貨日",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "is_valid",
                        "in": "query",
                        "description": "是否有效",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "uuid": {
                                            "type": "string",
                                            "example": "0b422f02-5acf-4bbb-bddf-4f6fdd843b08"
                                        },
                                        "inventory_no": {
                                            "type": "string",
                                            "example": "INV001"
                                        },
                                        "inventory_nm": {
                                            "type": "string",
                                            "example": "庫別1"
                                        },
                                        "inventory_qty": {
                                            "type": "integer",
                                            "example": "1000"
                                        },
                                        "lot_num": {
                                            "type": "string",
                                            "example": "LOT123"
                                        },
                                        "safety_stock": {
                                            "type": "integer",
                                            "example": "500"
                                        },
                                        "lastStock_receiptdate": {
                                            "type": "string",
                                            "example": "2025-03-31"
                                        },
                                        "is_valid": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "create_user": {
                                            "type": "string",
                                            "example": "admin"
                                        },
                                        "create_time": {
                                            "type": "string",
                                            "example": "admin"
                                        },
                                        "update_user": {
                                            "type": "string",
                                            "example": "2025-03-31T08:58:52.001975Z"
                                        },
                                        "update_time": {
                                            "type": "string",
                                            "example": "2025-03-31T08:58:52.001986Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "建立庫別失敗"
                    }
                }
            }
        },
        "/api/inventory/{inventoryno}": {
            "get": {
                "tags": [
                    "base_inventory"
                ],
                "summary": "查詢特定庫別資訊",
                "description": "查詢特定庫別資訊",
                "operationId": "getinventory",
                "parameters": [
                    {
                        "name": "inventoryno",
                        "in": "path",
                        "description": "庫別代號",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "uuid": {
                                            "type": "string",
                                            "example": "0b422f02-5acf-4bbb-bddf-4f6fdd843b08"
                                        },
                                        "inventory_no": {
                                            "type": "string",
                                            "example": "INV001"
                                        },
                                        "inventory_nm": {
                                            "type": "string",
                                            "example": "庫別1"
                                        },
                                        "inventory_qty": {
                                            "type": "integer",
                                            "example": "1000"
                                        },
                                        "lot_num": {
                                            "type": "string",
                                            "example": "LOT123"
                                        },
                                        "safety_stock": {
                                            "type": "integer",
                                            "example": "500"
                                        },
                                        "lastStock_receiptdate": {
                                            "type": "string",
                                            "example": "2025-03-31"
                                        },
                                        "is_valid": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "create_user": {
                                            "type": "string",
                                            "example": "admin"
                                        },
                                        "create_time": {
                                            "type": "string",
                                            "example": "admin"
                                        },
                                        "update_user": {
                                            "type": "string",
                                            "example": "2025-03-31T08:58:52.001975Z"
                                        },
                                        "update_time": {
                                            "type": "string",
                                            "example": "2025-03-31T08:58:52.001986Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "未找到庫別"
                    }
                }
            }
        },
        "/api/inventory2/{inventorynm}": {
            "get": {
                "tags": [
                    "base_inventory"
                ],
                "summary": "查詢特定庫別資訊",
                "description": "查詢特定庫別資訊",
                "operationId": "getinventorynm",
                "parameters": [
                    {
                        "name": "inventorynm",
                        "in": "path",
                        "description": "庫別名稱",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "uuid": {
                                            "type": "string",
                                            "example": "0b422f02-5acf-4bbb-bddf-4f6fdd843b08"
                                        },
                                        "inventory_no": {
                                            "type": "string",
                                            "example": "INV001"
                                        },
                                        "inventory_nm": {
                                            "type": "string",
                                            "example": "庫別1"
                                        },
                                        "inventory_qty": {
                                            "type": "integer",
                                            "example": "1000"
                                        },
                                        "lot_num": {
                                            "type": "string",
                                            "example": "LOT123"
                                        },
                                        "safety_stock": {
                                            "type": "integer",
                                            "example": "500"
                                        },
                                        "lastStock_receiptdate": {
                                            "type": "string",
                                            "example": "2025-03-31"
                                        },
                                        "is_valid": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "create_user": {
                                            "type": "string",
                                            "example": "admin"
                                        },
                                        "create_time": {
                                            "type": "string",
                                            "example": "admin"
                                        },
                                        "update_user": {
                                            "type": "string",
                                            "example": "2025-03-31T08:58:52.001975Z"
                                        },
                                        "update_time": {
                                            "type": "string",
                                            "example": "2025-03-31T08:58:52.001986Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "未找到庫別"
                    }
                }
            }
        },
...
    },
...
}